# contract/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chalk API contract preview",
    "version": "0.0.0-preview",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "https://api.chalk.q9labs.com",
      "description": "Production API"
    }
  ],
  "paths": {
...
    "/v1/telemetry/journey-events": {
      "post": {
        "operationId": "intakeJourneyEvents",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JourneyEventBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyEventIntake"
                }
              }
            },
            "description": "Accepted",
            "headers": {
              "x-chalk-journey-id": {
                "required": false,
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request",
            "x-chalk-error-codes": ["invalid_journey_event", "invalid_request"]
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized",
            "x-chalk-error-codes": ["unauthenticated"]
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large",
            "x-chalk-error-codes": ["payload_too_large"]
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests",
            "headers": {
              "Retry-After": {
                "required": true,
                "schema": {
                  "type": "integer"
                }
              },
              "X-RateLimit-Limit": {
                "required": true,
                "schema": {
                  "type": "integer"
                }
              },
              "X-RateLimit-Remaining": {
                "required": true,
                "schema": {
                  "type": "integer"
                }
              }
            },
            "x-chalk-error-codes": ["rate_limited"]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error",
            "x-chalk-error-codes": ["internal_error"]
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable",
            "x-chalk-error-codes": ["journey_ledger_unavailable", "service_unavailable"]
          }
        },
        "security": [
          {
            "sessionOrBearer": []
          }
        ],
        "summary": "Intake journey events",
        "x-chalk-max-body-bytes": 1048576,
        "x-chalk-rate-limit": {
          "limit": 600,
          "name": "v1.telemetry.intake",
          "window_seconds": 60
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionOrBearer": {
        "description": "Preview placeholder for routes accepted by Chalk session auth or bearer/API-key auth.",
        "scheme": "bearer",
        "type": "http"
      }
    },
    "schemas": {
...
      "DateTimeString": {
        "format": "date-time",
        "type": "string",
        "x-chalk-brand": "DateTimeString"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": ["code", "message"],
            "type": "object"
          }
        },
        "required": ["error"],
        "type": "object"
      },
...
      "JourneyEventBatch": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "attributes": {
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "maxLength": 1024,
                        "type": "string"
                      },
                      {
                        "type": "number"
                      },
                      {
                        "type": "boolean"
                      }
                    ]
                  },
                  "maxProperties": 32,
                  "type": "object"
                },
                "event_id": {
                  "$ref": "#/components/schemas/UUID"
                },
                "first_observed_layer": {
                  "minLength": 1,
                  "type": "string"
                },
                "journey_id": {
                  "$ref": "#/components/schemas/UUID"
                },
                "name": {
                  "minLength": 1,
                  "type": "string"
                },
                "occurred_at": {
                  "$ref": "#/components/schemas/DateTimeString"
                },
                "origin_kind": {
                  "minLength": 1,
                  "type": "string"
                },
                "parent_event_id": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/UUID"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "phase": {
                  "minLength": 1,
                  "type": "string"
                },
                "sequence": {
                  "type": "integer"
                },
                "span_id": {
                  "maxLength": 16,
                  "minLength": 16,
                  "pattern": "^(?=.*[1-9a-fA-F])[0-9a-fA-F]{16}$",
                  "type": ["string", "null"]
                },
                "state": {
                  "minLength": 1,
                  "type": "string"
                },
                "trace_id": {
                  "maxLength": 32,
                  "minLength": 32,
                  "pattern": "^(?=.*[1-9a-fA-F])[0-9a-fA-F]{32}$",
                  "type": ["string", "null"]
                },
                "upstream_visibility": {
                  "minLength": 1,
                  "type": "string"
                }
              },
              "required": ["event_id", "first_observed_layer", "journey_id", "name", "occurred_at", "origin_kind", "phase", "sequence", "state", "upstream_visibility"],
              "type": "object"
            },
            "maxItems": 100,
            "type": "array"
          }
        },
        "required": ["events"],
        "type": "object"
      },
      "JourneyEventIntake": {
        "additionalProperties": false,
        "properties": {
          "accepted_count": {
            "type": "integer"
          },
          "duplicate_count": {
            "type": "integer"
          },
          "journey_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": ["accepted_count", "duplicate_count", "journey_ids"],
        "type": "object"
      },
...
      "UUID": {
        "format": "uuid",
        "maxLength": 36,
        "minLength": 36,
        "type": "string",
        "x-chalk-brand": "UUID"
      },
...
    }
  }
}